+++ operate/client/src/modules/api/v2/messageSubscriptions/searchMessageSubscriptions.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+
+import {
+  endpoints,
+  type QueryMessageSubscriptionsRequestBody,
+  type QueryMessageSubscriptionsResponseBody,
+} from '@camunda/camunda-api-zod-schemas/8.8';
+import {requestWithThrow} from 'modules/request';
+
+const searchMessageSubscriptions = async (
+  payload: QueryMessageSubscriptionsRequestBody,
+) => {
+  return requestWithThrow<QueryMessageSubscriptionsResponseBody>({
+    url: endpoints.queryMessageSubscriptions.getUrl(),
+    method: endpoints.queryMessageSubscriptions.method,
+    body: payload,
+  });
+};
+
+export {searchMessageSubscriptions};